+++ CFDG.ACAD/CFDG.ACAD.csproj
@@ -113,14 +113,19 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CommandClasses\Calculations\CreateACPoints.cs" />
+    <Compile Include="CommandClasses\Calculations\GroupPoints.cs" />
+    <Compile Include="Extensions\PointGroupExtensions.cs" />
+    <Compile Include="Interfaces\ICommandMethod.cs" />
+    <Compile Include="CommandClasses\Calculations\MeasureDown.cs" />
     <Compile Include="Ribbon\Ribbon.cs" />
-    <Compile Include="TabCommands\Calculations.cs" />
+    <Compile Include="CommandClasses\Calculations.cs" />
     <Compile Include="Functions.cs" />
-    <Compile Include="TabCommands\Calculations\Footprint.cs" />
-    <Compile Include="TabCommands\ImportExport.cs" />
+    <Compile Include="CommandClasses\Calculations\Footprint.cs" />
+    <Compile Include="CommandClasses\Calculations\SlopeFromPoints.cs" />
+    <Compile Include="CommandClasses\ImportExport.cs" />
     <Compile Include="Main.cs" />
-    <Compile Include="TabCommands\PointGroupExtensions.cs" />
-    <Compile Include="TabCommands\ProjectManagement.cs" />
+    <Compile Include="CommandClasses\ProjectManagement.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -200,8 +205,6 @@
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Extensions\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
