+++ CFDG.ACAD/Extensions/PointGroupExtensions.cs
@@ -1,20 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Autodesk.AutoCAD.DatabaseServices;
 using Autodesk.Civil.ApplicationServices;
 using Autodesk.Civil.DatabaseServices;
 
-namespace CFDG.ACAD.TabCommands
+namespace CFDG.ACAD.Extensions
 {
     public static class PointGroupExtensions
     {
-        public static List<CogoPoint> GetCogoPoints(this PointGroup ptGroup)
+        public static List<CogoPoint> GetCogoPoints(this PointGroup pointGroup)
         {
             CivilDocument civDoc = CivilApplication.ActiveDocument;
-            return (from p in ptGroup.GetPointNumbers()
+            return (from p in pointGroup.GetPointNumbers()
                     select civDoc.CogoPoints.GetPointByPointNumber(p).GetObject(OpenMode.ForRead) as CogoPoint).ToList();
         }
     }
