+++ CFDG.ACAD/Interfaces/ICommandMethod.cs
@@ -1,0 +1,7 @@
+﻿namespace CFDG.ACAD
+{
+    public interface ICommandMethod
+    {
+        void InitialCommand();
+    }
+}
